+++ src/main/java/com/lesson/repository/tables/EduCourse.java
@@ -139,7 +139,7 @@ public class EduCourse extends TableImpl<EduCourseRecord> {
     /**
      * The column <code>lesson.edu_course.deleted</code>. 是否删除：0-未删除，1-已删除
      */
-    public final TableField<EduCourseRecord, Byte> DELETED = createField(DSL.name("deleted"), SQLDataType.TINYINT.nullable(false).defaultValue(DSL.inline("0", SQLDataType.TINYINT)), this, "是否删除：0-未删除，1-已删除");
+    public final TableField<EduCourseRecord, Integer> DELETED = createField(DSL.name("deleted"), SQLDataType.INTEGER.nullable(false).defaultValue(DSL.inline("0", SQLDataType.INTEGER)), this, "是否删除：0-未删除，1-已删除");
 
     private EduCourse(Name alias, Table<EduCourseRecord> aliased) {
         this(alias, aliased, null);
@@ -225,7 +225,7 @@ public class EduCourse extends TableImpl<EduCourseRecord> {
     // -------------------------------------------------------------------------
 
     @Override
-    public Row18<String, String, String, String, BigDecimal, BigDecimal, BigDecimal, BigDecimal, String, String, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Byte> fieldsRow() {
+    public Row18<String, String, String, String, BigDecimal, BigDecimal, BigDecimal, BigDecimal, String, String, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Integer> fieldsRow() {
         return (Row18) super.fieldsRow();
     }
 }
