+++ src/main/java/com/lesson/repository/tables/EduStudentCourseOperation.java
@@ -174,7 +174,7 @@ public class EduStudentCourseOperation extends TableImpl<EduStudentCourseOperati
     /**
      * The column <code>lesson.edu_student_course_operation.deleted</code>. 是否删除：0-未删除，1-已删除
      */
-    public final TableField<EduStudentCourseOperationRecord, Byte> DELETED = createField(DSL.name("deleted"), SQLDataType.TINYINT.nullable(false).defaultValue(DSL.inline("0", SQLDataType.TINYINT)), this, "是否删除：0-未删除，1-已删除");
+    public final TableField<EduStudentCourseOperationRecord, Integer> DELETED = createField(DSL.name("deleted"), SQLDataType.INTEGER.nullable(false).defaultValue(DSL.inline("0", SQLDataType.INTEGER)), this, "是否删除：0-未删除，1-已删除");
 
     private EduStudentCourseOperation(Name alias, Table<EduStudentCourseOperationRecord> aliased) {
         this(alias, aliased, null);
