+++ src/main/java/com/lesson/repository/tables/EduStudentCourseRecord.java
@@ -137,7 +137,7 @@ public class EduStudentCourseRecord extends TableImpl<EduStudentCourseRecordReco
     /**
      * The column <code>lesson.edu_student_course_record.deleted</code>. 是否删除：0-未删除，1-已删除
      */
-    public final TableField<EduStudentCourseRecordRecord, Byte> DELETED = createField(DSL.name("deleted"), SQLDataType.TINYINT.nullable(false).defaultValue(DSL.inline("0", SQLDataType.TINYINT)), this, "是否删除：0-未删除，1-已删除");
+    public final TableField<EduStudentCourseRecordRecord, Integer> DELETED = createField(DSL.name("deleted"), SQLDataType.INTEGER.nullable(false).defaultValue(DSL.inline("0", SQLDataType.INTEGER)), this, "是否删除：0-未删除，1-已删除");
 
     private EduStudentCourseRecord(Name alias, Table<EduStudentCourseRecordRecord> aliased) {
         this(alias, aliased, null);
@@ -228,7 +228,7 @@ public class EduStudentCourseRecord extends TableImpl<EduStudentCourseRecordReco
     // -------------------------------------------------------------------------
 
     @Override
-    public Row17<Long, String, String, String, String, LocalDate, LocalTime, LocalTime, BigDecimal, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Byte> fieldsRow() {
+    public Row17<Long, String, String, String, String, LocalDate, LocalTime, LocalTime, BigDecimal, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Integer> fieldsRow() {
         return (Row17) super.fieldsRow();
     }
 }
