+++ src/main/java/com/lesson/repository/tables/SysRole.java
@@ -69,7 +69,7 @@ public class SysRole extends TableImpl<SysRoleRecord> {
     /**
      * The column <code>lesson.sys_role.status</code>. 状态：0-禁用，1-启用
      */
-    public final TableField<SysRoleRecord, Byte> STATUS = createField(DSL.name("status"), SQLDataType.TINYINT.nullable(false).defaultValue(DSL.inline("1", SQLDataType.TINYINT)), this, "状态：0-禁用，1-启用");
+    public final TableField<SysRoleRecord, Integer> STATUS = createField(DSL.name("status"), SQLDataType.INTEGER.nullable(false).defaultValue(DSL.inline("1", SQLDataType.INTEGER)), this, "状态：0-禁用，1-启用");
 
     /**
      * The column <code>lesson.sys_role.created_time</code>. 创建时间
@@ -84,7 +84,7 @@ public class SysRole extends TableImpl<SysRoleRecord> {
     /**
      * The column <code>lesson.sys_role.deleted</code>. 是否删除：0-未删除，1-已删除
      */
-    public final TableField<SysRoleRecord, Byte> DELETED = createField(DSL.name("deleted"), SQLDataType.TINYINT.nullable(false).defaultValue(DSL.inline("0", SQLDataType.TINYINT)), this, "是否删除：0-未删除，1-已删除");
+    public final TableField<SysRoleRecord, Integer> DELETED = createField(DSL.name("deleted"), SQLDataType.INTEGER.nullable(false).defaultValue(DSL.inline("0", SQLDataType.INTEGER)), this, "是否删除：0-未删除，1-已删除");
 
     private SysRole(Name alias, Table<SysRoleRecord> aliased) {
         this(alias, aliased, null);
@@ -175,7 +175,7 @@ public class SysRole extends TableImpl<SysRoleRecord> {
     // -------------------------------------------------------------------------
 
     @Override
-    public Row7<Long, String, String, Byte, LocalDateTime, LocalDateTime, Byte> fieldsRow() {
+    public Row7<Long, String, String, Integer, LocalDateTime, LocalDateTime, Integer> fieldsRow() {
         return (Row7) super.fieldsRow();
     }
 }
